visit - www.hashmi.in and www.hashmidawakhana.in
Spermatorrhea can’t be considered a disease, but mostly a condition and it implies excessive ejaculation. This happens involuntary and in an uncontrollable manner. It is very close to what we call night emissions or wet dreams, but it is however slightly different. It does not occur only when a person is asleep and it’s not always triggered by erotic dreams or thoughts. Many young boys in their teenage years or in their early adulthood are scared after they experience night emissions, because they are worried that something bad is happening with them and they are suffering of a disease. In fact, sporadic ejaculations during sleep are nothing to worry about and they don’t signalize any serious medical attention. Unfortunately, the problem of involuntary ejaculations is not very much discussed with physicians because most men decide to keep it for themselves. They feel embarrassed about talking such intimate issues with their doctors and no one really knows for sure when night emissions turn into spermatorrhea.
If you feel any type of sexual health problem, first of all, talk frankly with your partner, do not hesitate, these remedies are tried and tested and Dr. Hashmi is known to help a good. 
Write or call us at Hashmi Dawakhana, Amroha, and Uttar Pradesh for better assessment.
For more information visit - www.hashmi.in and www.hashmidawakhana.in